% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantiles.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculates the quantile values of every numeric distribution in a rule data frame.}
\usage{
calculate_quantiles(rules)
}
\arguments{
\item{rules}{A data frame that contains a list column named \code{feature_conditions} and \code{distribution_values}.}
}
\value{
A data frame that contains the feature conditions in \code{rules} and their respective quantile values based on \code{distribution_values}.
}
\description{
This function calculates the quantile values (lower-whisker, first-quartile, median, third-quartile and upper-whisker) of every numeric distribution in the rule data frame.
It outputs a data frame composed by the feature conditions and the quantile values, all in separate columns.
}
\examples{
calculate_quantiles(rules)
}
