% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_rules.R
\name{caren_distribution_as_vector}
\alias{caren_distribution_as_vector}
\title{Converts the distribution outputted from Caren to a numeric vector.}
\usage{
caren_distribution_as_vector(caren_distribution)
}
\arguments{
\item{caren_distribution}{A string that follows the pattern: "error={value/count, ...}".}
}
\value{
A numeric vector that replicates each value by its count.
}
\description{
This function converts the distribution outputted from Caren, which is a string, to a numeric vector.
Note that this functions is a variation of function from \href{https://www.dcc.fc.up.pt/~amjorge/software/carenR/R/caren_integrated.R}{CarenR}.
}
\examples{
caren_distribution_as_vector("error={1/2,3/1}")
}
