% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_factual_subgroups.R
\name{counter_factual_subgroups}
\alias{counter_factual_subgroups}
\title{Calculates the counter factual subgroups of every subgroup in a rule data frame.}
\usage{
counter_factual_subgroups(rules, reference_quantiles)
}
\arguments{
\item{rules}{A data frame that contains a list column named \code{feature_conditions} and quantile values in the the following
columns: (\code{lw}, \code{q1}, \code{median}, \code{q3} and \code{uw}). This data frame can be obtained from the application of function
\code{edr::calculate_quantiles}.}

\item{reference_quantiles}{A vector that contains five float values that characterize the reference quantiles
(lower-whisker, first-quartile, median, third-quartile and upper-whisker).}
}
\value{
A object list, in which each object is composed of its feature conditions in \code{feature_conditions},
performance metrics in comparison to the reference values in \code{performance} and a data frame of counter factual subgroups.
}
\description{
This function calculates the counter factual subgroups of every subgroup in the rule data frame.
Taking an initial subgroup, a counter factual subgroup is one with a different set of feature conditions,
sharing at least one condition with the initial subgroup, but has an extreme opposite behavior.
It outputs a object list, in which each object is composed of its feature conditions in \code{feature_conditions},
performance metrics in comparison to the reference values in \code{performance} and a data frame of variation subgroups
that have contrasting performance while sharing at least one feature condition.
}
\examples{
counter_factual_subgroups(example_rules \%>\% calculate_quantiles() \%>\% head(50), c(0.00742, 0.547, 0.926, 1.33, 2.42))
}
