% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_graph}
\alias{create_graph}
\title{Creates two data frames, one composed of edges and other of vertices that form a graph based on a set of rules.}
\usage{
create_graph(rules, reference_value)
}
\arguments{
\item{rules}{A data frame of rules composed by \code{feature_conditions}, \code{antecedent_support}, \code{p_value}, \code{kurtosis}, \code{skewness}, \code{mean}, \code{mode}, \code{median} and \code{standard_deviation}.}

\item{reference_value}{A numeric value used to color the subgroup vertices based on their medians.}
}
\value{
A named list of two data frames: \code{vertices}, being composed of all graph vertices, and \code{edges}, containing all connections between vertices.
}
\description{
This function creates two data frames. One is composed of all the edges that define the graph, in the form \code{from}-\code{to}.
The vertex data frame is more detailed, containing identifiers, labels, shapes, colors, etc..
The color of each vertex is calculated based on its median compares to the input \code{reference_value}.
}
\examples{
create_graph(example_rules, 0.5)
}
\seealso{
[edr::plot_graph()] to understand how the output of this function is used graphically.
}
