% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwr-package.R
\docType{package}
\name{edr-package}
\alias{edr}
\alias{edr-package}
\title{edr: edr}
\description{

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (0000-0003-3925-190X)

}
\keyword{internal}
