% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_rules.R
\name{filter_subgroups}
\alias{filter_subgroups}
\title{Filters data frame based on a regular expression applied to the feature conditions that define a subgroup.}
\usage{
filter_subgroups(rules, regex = "")
}
\arguments{
\item{rules}{A data frame that contains a list column named \code{feature_conditions}.}

\item{regex}{A regular expression to filter the feature conditions in the input data frame.}
}
\value{
A vector of strings representing the feature conditions that define the subgroups in the input data frame.
}
\description{
This function filters out all data items that do not match the input regular expression.
}
\examples{
filter_subgroups(rules, regex = "lstat=]")
}
