% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_rules.R
\name{get_subgroups}
\alias{get_subgroups}
\title{List all subgroups in an orderly manner.}
\usage{
get_subgroups(rules)
}
\arguments{
\item{rules}{A data frame that contains a list column named \code{feature_conditions}.}
}
\value{
A vector of strings representing the feature conditions that define the subgroups in the input data frame.
}
\description{
This function returns a vector of strings, each representing the feature conditions that define a subgroup.
}
\examples{
get_subgroups(rules)
}
