% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_boxplots.R
\name{performance_boxplots}
\alias{performance_boxplots}
\title{Plots the distributions of input rules using boxplots and colors their quantile values based on how these compare to a reference distribution.}
\usage{
performance_boxplots(
  rules,
  reference_conditions,
  reference_distribution,
  label = ""
)
}
\arguments{
\item{rules}{A data frame of rules composed by \code{feature_conditions}, \code{antecedent_support}, \code{p_value}, \code{kurtosis}, \code{skewness}, \code{mean}, \code{mode}, \code{median} and \code{standard_deviation}.}

\item{reference_conditions}{A list of strings that defines the feature conditions of the reference.}

\item{reference_distribution}{A numeric vector composed of the distribution values of the reference.}

\item{label}{A string that represents the label used on the distribution axis.}
}
\value{
A \code{ggplot2} object composed of a canvas of boxplots from the rules.
}
\description{
This function generates one canvas with multiple boxplots of the rules, each with three possible colors for each quantile: red, blue or green.
The first color means that the distribution of the subgroup for that quantile has a higher value than the reference for the same quantile, blue the same and green a lower value.
It uses \code{ggplot2} to produce the plot objects.
}
\examples{
performance_boxplots(rules = t \%>\% head(-1), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), label = "error")
performance_boxplots(rules = t \%>\% head(40), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), label = "error")
}
