% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plots an interactive graph based on a data frame of vertices and a data frame of edges.}
\usage{
plot_graph(vertices, edges)
}
\arguments{
\item{vertices}{A data frame the contains \code{from} and \code{to} columns.}

\item{edges}{A data frame the contains \code{id}, \code{label}, \code{group}, \code{value}, \code{shape}, \code{title}, \code{color} and \code{shadow} columns.}
}
\value{
An interactive \code{visNetwork} graph of input rules.
}
\description{
This function creates an interactive visualization of the rules as a graph.
It uses the library \code{visNetwork} for the graphical component.
}
\examples{
graph_data <- example_rules \%>\% create_graph(0.5)
plot_graph(graph_data$vertices, graph_data$edges)
}
\seealso{
[edr::create_graph()] to understand how the inputs of this function are calculated.
}
