% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_single_subgroup.R
\name{plot_grid_single_subgroup}
\alias{plot_grid_single_subgroup}
\title{Creates a grid of plots that compare pairs of distributions, with one being the reference.}
\usage{
plot_grid_single_subgroup(
  rules,
  reference_conditions,
  reference_distribution,
  page = 1,
  label = "error",
  type = "Boxplot",
  items_per_page = 4
)
}
\arguments{
\item{rules}{A data frame of rules composed by \code{feature_conditions} and \code{distribution_values}.}

\item{reference_conditions}{A list of strings that defines the feature conditions of the reference.}

\item{reference_distribution}{A numeric vector composed of the distribution values of the reference.}

\item{page}{A numeric value indicating the page, i.e. the subset of input rules, that are to be displayed.}

\item{label}{A string that represents the label used on the distribution axis.}

\item{type}{A string that contains the type of plot to be produced. It can be "boxplot", "density", "dotplot", "histogram", "scatterplot" or "violinplot".}

\item{items_per_page}{a numeric value that indicates how many plots should be displayed in a single grid.}
}
\value{
A grid of \code{ggplot2} objects, each portraying a subgroup and the reference.
}
\description{
This function generates a grid of plots, each with two distributions, one representing the subgroup and another representing the reference.
It uses \code{ggplot2} to produce the plot objects.
}
\examples{
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "boxplot", label = "error", items_per_page = 5)
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "density", label = "error", items_per_page = 9)
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "dotplot", label = "error", items_per_page = 4)
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "histogram")
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "scatterplot", label = "error", items_per_page = 4)
plot_grid_single_subgroup(t, reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), type = "violinplot", label = "error", items_per_page = 4)
}
