% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{plot_multiple_models}
\alias{plot_multiple_models}
\title{Plots the distributions of input rules grouped by and colored by their respective models.}
\usage{
plot_multiple_models(rules, title = "", label = "error", type = "Boxplot")
}
\arguments{
\item{rules}{A data frame of rules composed by \code{feature_conditions}, \code{antecedent_support}, \code{p_value}, \code{kurtosis}, \code{skewness}, \code{mean}, \code{mode}, \code{median} and \code{standard_deviation}.}

\item{title}{A string used as the title of the canvas.}

\item{label}{A string that represents the label used on the distribution axis.}

\item{type}{A string that contains the type of plot to be produced. It can be "boxplot", "density", "dotplot", "histogram", "scatterplot" or "violinplot".}
}
\value{
A \code{ggplot2} object composed of the plots of the rules divided by subgroup and colored by model.
}
\description{
This function generates one or more plots of the rules divided by subgroup and colored by model.
It uses \code{ggplot2} to produce the plot objects.
}
\examples{
rule_1 <- t \%>\% head(1) \%>\% mutate(model = "SVM")
rule_2 <- rule_1 \%>\% mutate(distribution_values = (distribution_values \%>\% unlist(use.names = FALSE) + runif(distribution_values \%>\% unlist(use.names = FALSE) \%>\% length(), -1, 1)) \%>\% list(), model = "ANN")
reference_1 <- t \%>\% tail(1) \%>\% mutate(model = "SVM")
reference_2 <- reference_1 \%>\% mutate(distribution_values = (distribution_values \%>\% unlist(use.names = FALSE) + runif(distribution_values \%>\% unlist(use.names = FALSE) \%>\% length(), -1, 1)) \%>\% list(), model = "ANN")

plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "boxPlot")
plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "DENSITY")
plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "dotplot")
plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "hisTOgRAM")
plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "SCATTERplot")
plot_multiple_models(rules = rbind(rule_1, rule_2), title = "Example of a title", label = "errors", type = "violinplot")

plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "boxPlot")
plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "DENSITY")
plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "dotplot")
plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "hisTOgRAM")
plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "Scatterplot")
plot_multiple_models(rules = rbind(rule_1, rule_2, reference_1, reference_2), title = "Example of a title", label = "errors", type = "Violinplot")
}
