% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{plot_multiple_subgroups}
\alias{plot_multiple_subgroups}
\title{Plots the distributions of the input subgroups and the reference.}
\usage{
plot_multiple_subgroups(
  rules,
  reference_conditions,
  reference_distribution,
  title = "",
  label = "error",
  type = "Boxplot"
)
}
\arguments{
\item{rules}{A data frame of rules composed by \code{feature_conditions} and \code{distribution_values}.}

\item{reference_conditions}{A list of strings that defines the feature conditions of the reference.}

\item{reference_distribution}{A numeric vector composed of the distribution values of the reference.}

\item{title}{A string used as the title of the canvas.}

\item{label}{A string that represents the label used on the distribution axis.}

\item{type}{A string that contains the type of plot to be produced. It can be "boxplot", "density", "dotplot", "histogram", "scatterplot" or "violinplot".}
}
\value{
A \code{ggplot2} object composed of a canvas portraying the input subgroups and the reference.
}
\description{
This function generates one canvas with multiple distributions representing the input subgroups and the reference.
It uses \code{ggplot2} to produce the plot objects.
}
\examples{
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "BOXPLOT")
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "density")
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "dotplot")
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "Histogram")
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "Scatterplot")
plot_multiple_subgroups(rules = t \%>\% head(5), reference_conditions = t \%>\% tail(1) \%>\% pull(feature_conditions), reference_distribution = t \%>\% tail(1) \%>\% pull(distribution_values) \%>\% unlist(use.names = FALSE), title = "First 5 Subgroups", label = "error", type = "ViolinPlot")
}
