% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{plot_single_subgroup}
\alias{plot_single_subgroup}
\title{Plots the distributions of the input subgroup and reference.}
\usage{
plot_single_subgroup(
  subgroup_conditions,
  subgroup_distribution,
  reference_conditions,
  reference_distribution,
  type = "Boxplot",
  label = "error"
)
}
\arguments{
\item{subgroup_conditions}{A list of strings that defines the feature conditions of the subgroup.}

\item{subgroup_distribution}{A numeric vector composed of the distribution values of the subgroup.}

\item{reference_conditions}{A list of strings that defines the feature conditions of the reference.}

\item{reference_distribution}{A numeric vector composed of the distribution values of the reference.}

\item{type}{A string that contains the type of plot to be produced. It can be "boxplot", "density", "dotplot", "histogram", "scatterplot" or "violinplot".}

\item{label}{A string that represents the label used on the distribution axis.}
}
\value{
A \code{ggplot2} object composed of a canvas portraying the subgroup and the reference.
}
\description{
This function generates one canvas with two distributions, one representing the subgroup and another representing the reference.
It uses \code{ggplot2} to produce the plot objects.
}
\examples{
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "Boxplot", label = "error")
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "density", label = "error")
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "dotplot", label = "error")
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "histogram", label = "error")
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "scatterplot", label = "error")
plot_single_subgroup(subgroup_conditions = t[1,]$feature_conditions, subgroup_distribution = t[1,]$distribution_values \%>\% unlist(use.names = FALSE), reference_conditions = t[nrow(t),]$feature_conditions, reference_distribution = t[nrow(t),]$distribution_values \%>\% unlist(use.names = FALSE), type = "violinplot", label = "error")
}
