% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwr-package.R
\docType{package}
\name{rwr-package}
\alias{rwr}
\alias{rwr-package}
\title{rwr: rwr}
\description{

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (0000-0003-3925-190X)

}
\keyword{internal}
