% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{single_model_data}
\alias{single_model_data}
\title{Calculates the prediction errors of a model and adds it to the input data frame}
\usage{
single_model_data(
  df,
  predicted,
  feature_y,
  type = "Absolute",
  output_predictions = FALSE,
  output_target = FALSE
)
}
\arguments{
\item{df}{A data frame that contains feature values and a target column.}

\item{predicted}{A vector that contains the predicted values by the model.}

\item{feature_y}{A string that contains the name of the column that represents the target variable.}

\item{type}{A string that contains the type of error to be calculated. It can be "absolute", "residual" or "logarithmic".}

\item{output_predictions}{A boolean that defines if the output data frame will contain the predictions in a new column.}

\item{output_target}{A boolean that defines if the output data frame will contain the target values.}
}
\value{
A data frame that contains the at least one extra column in the form of prediction errors based on the type of error selected.
}
\description{
This function creates a data frame with feature values, the model's errors and predictions.
Note that this functions is a variation of function from \href{https://github.com/inesareosa/MScThesis}{Error Dependence Plots}.
}
\examples{
single_model_data(df = iris, iris$Sepal.Length + runif(length(iris$Sepal.Length), -1, 1), feature_y = "Sepal.Length", type = "Logarithmic")

single_model_data(df = iris, iris$Sepal.Length + runif(length(iris$Sepal.Length), -1, 1), feature_y = "Sepal.Length", type = "Logarithmic", output_predictions = FALSE, output_target = FALSE)

single_model_data(df = iris, iris$Sepal.Length + runif(length(iris$Sepal.Length), -1, 1), feature_y = "Sepal.Length", type = "Absolute", output_predictions = FALSE, output_target = TRUE)

single_model_data(df = iris, iris$Sepal.Length + runif(length(iris$Sepal.Length), -1, 1), feature_y = "Sepal.Length", type = "Residual", output_predictions = TRUE, output_target = FALSE)

single_model_data(df = iris, iris$Sepal.Length + runif(length(iris$Sepal.Length), -1, 1), feature_y = "Sepal.Length", type = "Residual", output_predictions = TRUE, output_target = TRUE)
}
\seealso{
[calculate_error()] to understand how error values are calculated.
}
