% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_covid_cases}
\alias{uk_covid_cases}
\title{World Health Organization TB data}
\format{
\subsection{\code{uk_covid_cases}}{

A data frame with 436 rows and 2 columns:
\describe{
\item{reporting_date}{Date cases were reported}
\item{new}{Number of new daily confirmed cases}
}
}
}
\source{
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
uk_covid_cases
}
\description{
UK specific COVID-19 case data from the COVID-19 Data Repository
by the Center for Systems Science and Engineering (CSSE)
at Johns Hopkins University.
}
\details{
This data consists of newly confirmed daily cases in the UK from
the 22nd January 2020 to the 1st April 2021.
}
\keyword{datasets}
