% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr}
\alias{edr}
\title{Estimated Dissemination Ratio}
\usage{
edr(x, window = 1, na_rm = FALSE, simulations = 0, alpha = 0.05)
}
\arguments{
\item{x}{\verb{[numeric]} - A vector of case numbers.}

\item{window}{\verb{[integer(1)]} - An integer denoting the window size.}

\item{na_rm}{\verb{[logical(1)]} - Should \code{NA} values be ignored? Default is \code{FALSE}.}

\item{simulations}{\verb{[numeric(1)]} - Number of poisson simulations for
both the numerator and denominator. If simulations is 0 then the result is
a numeric vector, otherwise a matrix of the estimate, lower confidence interval
and upper confidence interval are returnd.}

\item{alpha}{\verb{[numeric(1)]} - Alpha significance level.
The default is a 95\% (1 - 0.05) confidence interval.}
}
\value{
A numeric vector of edr estimates when \code{simulations} is 0, otherwise
a 3-col matrix of estimates, lower and upper confidence intervals.
}
\description{
Estimated Dissemination Ratio
}
\details{
\subsection{Confidence intervals}{

To generate confidence intervals, simply specify
a \code{simulations} value greater than 0.
}
}
\references{
Perez-Reche FJ, Taylor N, McGuigan C, Conaglen P, Forbes K, Strachan N, Honhold N (2021) Estimated Dissemination Ratio --- A practical alternative to the reproduction number for infectious diseases. Frontiers in Public Health 9. DOI:  10.3389/fpubh.2021.675065.
}
