% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr}
\alias{edr}
\alias{edr_only}
\alias{print.edr}
\title{Estimated Dissemination Ratio}
\usage{
edr(
  x,
  window = 1,
  order_by = NULL,
  na_rm = FALSE,
  simulations = 0,
  alpha = 0.05
)

edr_only(x, window = 1, na_rm = FALSE)

\method{print}{edr}(x, ...)
}
\arguments{
\item{x}{\verb{[numeric]} - A vector of case numbers.}

\item{window}{\verb{[integer(1)]} - An integer denoting the window size. \cr
Default is \code{1}.}

\item{order_by}{An optional time index (e.g. date vector) to use.
This will ensure EDR estimates are calculated in the correct order. It
will also be used in the x-axis when calling \code{plot.edr()}.}

\item{na_rm}{\verb{[logical(1)]} - Should \code{NA} values be ignored? \cr
Default is \code{FALSE}.}

\item{simulations}{\verb{[numeric(1)]} - Number of Poisson simulations for
both the numerator and denominator. \cr
If simulations is 0 (the default) then the result is
a numeric vector, otherwise a data frame of the estimate,
lower confidence interval and upper confidence interval
is returned.}

\item{alpha}{\verb{[numeric(1)]} - Alpha significance level. \cr
The default is a 95\% (1 - 0.05) confidence interval.}

\item{...}{Arguments passed onto \code{print.edr}.}
}
\value{
\code{edr()} returns a \code{data.table} of case counts and EDR estimates,
optionally along with a user-supplied time index, and confidence intervals. \cr
\code{edr_only()} returns \bold{only} the EDR estimates as a numeric vector. \cr
}
\description{
Estimated Dissemination Ratio
}
\details{
\subsection{Ratio}{

EDR is a simple ratio calculated as the number of cases in a time interval,
divided by the number of cases in a preceding time interval
of the same width.
Therefore it is easy to produce, both conceptually and computationally,
making it a useful metric in analysing epidemic curves.

For example, to calculate the 7-day EDR, the first ratio would be
the number of cases in week 2 (days 8 to 14),
divided by the number of cases in week 1 (days 1 to 7).
You then apply this on a rolling basis, e.g. the number of cases between
days 9 to 15 divided by the number of cases between days 2 to 8, and so on.

A ratio \bold{equal to} 1 signifies the cases have remained stable in
the current time period relative to the previous one. \cr
A ratio \bold{above} 1 signifies the cases have increased in
the current time period relative to the previous one.\cr
A ratio \bold{below} 1 signifies the cases have decreased in
the current time period relative to the previous one.
}

\subsection{Confidence intervals}{

Confidence intervals are calculated by assuming the cases across both
intervals are Poisson distributed random variables.
Random deviates are then generated for each and ratios are calculated
across these two sets. Percentile confidence intervals are produced
by taking the \code{alpha/2} and \code{1 - (alpha/2)} percentiles of the ratios.
}
}
\examples{
library(edr)
library(outbreaks)
library(data.table)
library(ggplot2)

ebola <- as.data.table(ebola_sim_clean$linelist)
cases <- ebola[, .(n = .N), keyby = date_of_onset]

edr <- cases[, edr(n, order_by = date_of_onset, window = 7)]

plot(edr) +
  geom_smooth(se = FALSE)
}
\references{
Perez-Reche FJ, Taylor N, McGuigan C, Conaglen P, Forbes K, Strachan N,
Honhold N (2021) Estimated Dissemination Ratio ---
A practical alternative to the reproduction number for infectious diseases.
Frontiers in Public Health 9. DOI:  10.3389/fpubh.2021.675065.
}
