% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.edr}
\alias{plot.edr}
\title{Plot EDR}
\usage{
\method{plot}{edr}(x, include_cases = TRUE, ...)
}
\arguments{
\item{x}{\verb{[numeric]} - An object of class "edr",
created using the \code{edr()} function.}

\item{include_cases}{\verb{[logical(1)]} Should case counts be plotted alongside the EDR
estimates? The default is \code{TRUE}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} of the EDR trend, optionally alongside case counts.
}
\description{
A plot method for an object of class \code{edr} which is returned
by \code{edr()}.
}
\seealso{
\code{\link[=edr]{edr()}}
}
